+++ src/main/java/tripticket/location/dao/LocationDao.java
@@ -3,6 +3,6 @@ package tripticket.location.dao;
 import trioticket.location.model.Location;
 import tripticket.generic.dao.GenericDao;
 
-public class LocationDao extends GenericDao<Location, Long> implements LocationDaoI {
+public class LocationDao extends GenericDao<Location, Integer> implements LocationDaoI {
 	
 }
