+++ src/main/java/tripticket/location/dao/LocationDaoI.java
@@ -4,6 +4,6 @@ import trioticket.location.model.Location;
 import tripticket.generic.dao.GenericDaoI;
 
 
-public interface LocationDaoI extends GenericDaoI<Location, Long>{
+public interface LocationDaoI extends GenericDaoI<Location, Integer>{
 	
 }
